+++ src/constants.ts
@@ -295,6 +295,7 @@ export class Constants {
         GOTO_LINE_IN_LEO_OUTLINE: Constants.NAME + ".gotoLineInLeoOutline",
         IMPORT_INTO_LEO_OUTLINE: Constants.NAME + ".importIntoLeoOutline",
         HANDLE_UNL: Constants.NAME + ".handleUnl",
+        UNL_TO_CLIPBOARD: Constants.NAME + ".unlToClipboard",
         SHORT_GNX_UNL_TO_CLIPBOARD: Constants.NAME + ".shortGnxUnlToClipboard",
         FULL_GNX_UNL_TO_CLIPBOARD: Constants.NAME + ".fullGnxUnlToClipboard",
         SHORT_LEGACY_UNL_TO_CLIPBOARD: Constants.NAME + "shortLegacyUnlToClipboard",
@@ -428,6 +429,8 @@ export class Constants {
         DELETE_MARKED_NODES: Constants.NAME + ".deleteMarkedNodes",
         MOVE_MARKED_NODES: Constants.NAME + ".moveMarkedNodes",
 
+        FIND_HISTORY_BACK: Constants.NAME + ".findHistoryBack",
+        FIND_HISTORY_FORWARD: Constants.NAME + ".findHistoryForward",
         FIND_QUICK: Constants.NAME + ".findQuick",
         FIND_QUICK_SELECTED: Constants.NAME + ".findQuickSelected",
         FIND_QUICK_TIMELINE: Constants.NAME + ".findQuickTimeline",