+++ src/leo-ui.ts
@@ -30,6 +30,7 @@ import {
     QuickPickOptions,
     ReqRefresh,
     RevealType,
+    SearchSettingNames,
     UnlType
 } from "./types";
 import { StringTextWrapper } from "./core/leoFrame";
@@ -2098,6 +2099,13 @@ export class LeoUI extends NullGui {
         workspace.logPane.showTab('nav');
     }
     //@+node:felix.20260322235817.1: *3* Search and Replace
+    //@+node:felix.20260601163603.1: *4* doArrow
+    public doArrow(key: "Up" | "Down") {
+        const c = g.app.windowList[this.frameIndex].c;
+        const fc = c.findCommands;
+        fc.do_arrow(key);
+    }
+
     //@+node:felix.20260410231431.1: *4* startSearch
     /**
      * * Opens the find panel and selects all & focuses on the find field.
@@ -2312,9 +2320,28 @@ export class LeoUI extends NullGui {
             }
         }
     }
+    //@+node:felix.20260528002845.1: *4* newHeadline
+    /**
+     * * Set search setting in the search webview
+     * @param p_id string id of the setting name
+     */
+    public setSearchSetting(p_id: SearchSettingNames): void {
+        workspace.logPane.setSearchSetting(p_id);
+
+        // we need to replace this code from LeoJS to the new leo-web appropriate implementation. 
+        /*
+        if (this._findPanelWebviewExplorerView) {
+            void this._findPanelWebviewExplorerView!.webview.postMessage({ type: 'setSearchSetting', id: p_id });
+        }
+        if (this._findPanelWebviewView) {
+            void this._findPanelWebviewView!.webview.postMessage({ type: 'setSearchSetting', id: p_id });
+        }
+        */
+    }
+
     //@+node:felix.20260410231409.1: *4* loadSearchSettings
     /**
-     * * Gets the search settings from Leo, and applies them to the find panel webviews
+     * * Gets the search settings from Leo, and applies them to the find panel webview
      */
     public loadSearchSettings(): void {
 
@@ -2466,12 +2493,10 @@ export class LeoUI extends NullGui {
         const suboutlineOnly = searchSettings.suboutline_only || false;
 
         if (!nodeOnly && !suboutlineOnly && !fileOnly) {
-            find.entire_outline = true;
             if (!w.isChecked()) {
                 w.toggle();
             }
         } else {
-            find.entire_outline = false;
             if (w.isChecked()) {
                 w.toggle();
             }