+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## 1.0.22
 
-- ...
+- Added 'Copy UNL to clipboard' to the body and outline context-menus.
+- Added tooltips to the body pane's UNL and URL links.
+- Fixed handleUrlHelper to give focus to the opened browser tab in all circumstances.
 
 ## 1.0.21